(feat) Support blending transforms in the viewer
This is render opcode 0x09. It appears to store a linear combination
of products of stack matrices with static matrices from the last
unknown section of the model data onto the stack.

Roxas from Kingdom Hearts now loads and animates correctly!

The implementation for the joint tree and exporter is much more
complicated, so it will be in a separate commit.