(bug) Fix loss of precision in time calculations
I had assumed that precise_time_s used the program start as the
epoch, but the epoch is actually unspecified (boot time, maybe?).
This meant that I has fewer bits that I thought in time values.
The loss of significance caused slow and jerky animations/movement.
Techincally, the fix (which is to store time values as f64s and
cast to f32 after taking the difference) only delays the inevitable.
Since the epoch is unspecified, the values could _still_ be too
large to represent the fraction of a second we need.

Also, advance the animation multiple frames if necessary.