(feat) Remove dummy root if possible
This results in a cleaner armature for DCC tools.

Also, each <controller> only contains the joint data for nodes that
influence at least one of its vertices. There was also a general
cleanup to remove the assumption that the joint indices were all in
a continuous range 0..N.

I'm still interesting in only emitting the inverse bind matrices
once though. Revisit this... someday :)