(feat) Add extracting BMD/BCA files from archives
Where, by archive, I mean any file where the BMD or BCA is just
located at some non-zero offset in the file. This includes
uncompressed ROM images.

The reason for including extraction in this program (whose main
purpose is conversion to COLLADA) is that we already have the
functionality to attempt to parse a file and return an Err if we
fail, which means we can easily find any files we will be able to
read in.