(bug) Load palette colors lazily
Like how a 256-color palette image doesn't require the palette to
actually contain 256 colors if all those indices aren't used, we
now only require a compressed texture to contain all four colors
in a palette if the mode and texel demand that color be read.

If it does demand an OOB color be read, it will still crash however.
This should probably be fixed to allow loading of models with
corrupted textures.