(bug) Remove cur_stack_pos
This is from the original kiwi.ds code, I think? Anyway, apparently
the 0xX6 commands do not store to the stack unless the command
specifies a location to store to, so we don't need to track a current
position at all!

Fixes the last bugs I know of in Ore wa.

Actually, it's pretty amazing that this worked as well as did with
this bug!