(bug) Read inverse bind matrices lazily
This is similar to 7b18af2. A model doesn't need an inverse bind
matrix unless its index is actually used by the blending opcode.
There was probably a converter tool that turned a normal skeleton
into a BMD file with object matrices and inverse binds in a one-to-
one relationship, so there are typically num_objects inv. binds,
but there don't have to be. We now find and decode them lazily so
we only need as many as are actually used.

I don't like that Cur plays double duty as both a position in a
buffer and the base of a View of unknown length. Refactor this
when I have time...