(bug) Emit each <image> only once
Since one <image> tag was emitted for every material with a texture
and the name was used as the ID, materials with the same texture/
palette resulting in <image> elements with the same ID.

Use a HashSet to emit each only once now.

This did not, AFAIK, trigger any errors in any COLLADA readers
though.