(ref)(feat) Add simple decompression to extractor
If we find a stamp in a ROM but fail to parse a file there, we now
try doing LZ77 decompression there and see if we can parse the
resulting decompressed buffer.

Can now successfully extract >1000 models from Love Plus.

This is still WIP and needs to be cleaned up...