(feat) Be more robust about texture matrices
I think one of the unknown fields, and not the transform mode,
determines whether there is texture matrix data (changing the
transform mode doesn't cause the DS to change the texture matrix
in the debugger). This doesn't really fix anything, it just
removes the unimplemented!() so we can make progress.