Better handling of non-invertible matrices
Perturb singular object matrices in an attempt to make them
invertible. This also means the converter needs to use a set of
objects matrices distinct from the ones in the model file.

"Fixes" (I use the term loosely) an animation xTr sent me where
a scale of 0 is used to hide certain meshes until the point
in the animation where they're supposed to appear.