Build a table of which items to use for materials
Previosuly we would compute this on the fly from the name tables,
but now we compute it as part of building the DB. The main advantage
is we can prefer textures/palettes from the same file as a model.
You can now bulk convert all of 俺がオマエを守る and we won't get
confused because different models used the same name for eg. the
wall textures.

Since we compute materials in a single place, we now also emit some
help for #7 now.

This also required building tables of which file each Nitro item
was found in. info! is used as of this commit now too.