Add --apply-any-animation option
This disables our heuristic that the number of TRS curves in an
animation and the number of objects in a model must be the same.
If there too few TRS curves, we use the identity for the extra
object matrices.

This is needed for certain models in Ore ga Omae o Mamoru, eg.
the mecha boss.