Add Connection abstraction for recording interrelationships
Factor off the knowledge about what texture/palette to use for each
material (which is currently in the db.material_table) and what
animations apply (which is currently decided on the fly in both
the viewer and converter) into a struct.

Only the info module has been updated to use the Connection.
Others forthcoming.