Emit quads when possible in converter
Primitives was taught to record polys as quads, and the converter
uses it to give more accurate COLLADA files. Decomposing a quad
into tris gives interpolation artifacts, since interpolation should
depend on all four vertices but when decomposed it can depend on
only three. For a good description of this, see

https://www.supermodel3.com/Forum/viewtopic.php?f=7&t=1564

which is about the Dreamcast, but it applies to the DS too.

However, the *viewer* still uses tris. Implementing correct quad
rendering is not nearly worth it there :)