Better animation compatibility
Animations in Rune Factory failed to parse. Fixed by growing the
"is this TRS property animated" field to two bits. This uses the
previously-unknown bit between the is-animated and is-const fields,
and seems to work but this can't really be right, since there's no
reason to use a two-bit field for that...

We also check for start_frame < end_frame so we don't panic when
we try to subtract those, and we ensure num_frames > 0 since that's
assumed in lots of places that use it anyway.