Add a light to the viewer
For draw calls that use normals, use a new vertex shader that
calculates the vertex color in a DS-like way. Draw calls without
normals are unchanged.

This is done per-draw-call, when for accuracy it should be done
per-vertex, but this is likely to be okay in practice.

The vertex color is not exactly what it would be on the DS. The
specular term in ignored and I think the DS may have separate
matrices used to transform the normals. I used melonDS's
CalculateLighting as a reference for how the DS works.

Adrot's models look better in the viewer now.