Name duplicate extracted files like filename.XXX.nsbmd
Previously the XXX part was appended directly to the file name, but
that made it hard to tell if a number at the end of the file name was
part of the original filename or was appended because it was a duplicate.
The preceeding dot now disambiguates.

This also uses the whole filename+extension to determine uniqueness.
Previously there was a bug where foo.nsbmd and foo.nsbtx would be
considered the same and, eg. the latter would be renamed foo1.nsbtx.