CLI: write custom arg parsing code
This drops the clap dependency, which is a boost for compile time/
build size/binary size. Also alows us to use a consistent method
for reporting errors.

Error and usage messages got a little love too.

Edges are still a bit rough on this I think, but can work on it
more later.